# idb and webSQL adapters should only use one transaction for allDocs with keys (6854)