PR for #6854