@nename0 2017-11-21
I also rewrote the WebSql adapter to use the keys in a SQL "... WHERE id in (keys)" statement. This gave a moderate performance boost aswell.

Note: I will remove the deleted: 'ok' option from the allDocs method in WebSql and IDB. It's not documented and only used for the keys option. So I think it's ok to just remove it (?)