@nename0 2017-11-22
Results of new [all-docs-keys](https://github.com/pouchdb/pouchdb/pull/6858/commits/b6977fcf2a737dbea8bef5cf3673267d4cb0f602#diff-43963c040e192bd7d42b1bec99c2fc5bR138) performance test:

| Median  | Before | After |
| ------------- |:-------------:| -----:|
| Chrome 62 idb  | 32.3 ms | 4.5 ms |
| Firefox 57 idb | 12.3 ms | 3.5 ms |
| Chrome 62 websql| 33.4 ms | 1.7 ms |