@nename0 2017-11-22
#### Issuses: 

- [Build Job 32](https://travis-ci.org/pouchdb/pouchdb/jobs/305733391):
I traced down the reason for [this](https://travis-ci.org/pouchdb/pouchdb/jobs/305733391#L2870) error. It's because pouchdb-server does not use a updated/npm linked version of `pouchdb-adapter-leveldb-core`. Using a updated version of `bin/run-tests.sh` ([gist](https://gist.github.com/nename0/e78a8297aadef97b5808eb2665d37f83#file-run-tests-sh-L9)) it works. For some reason I also had to change the npm link commands to get it working.
So in order to fix this build you'd need to update the `bin/run-tests.sh` file, but it does not really fit in this PR