@daleharvey 2017-11-23
This looks great, thanks so much, I have restarted #15 and #25, its likely they are intermittent failures unrelated to your PR, as for #32 its great you found that problem as another PR had the exact same issue, I will open a new issue for that if you could submit your updated run-tests for that, then when that lands rebase this and should be all good to go? thanks