@daleharvey 2017-11-23
@nename0 https://github.com/pouchdb/pouchdb/issues/6859 for the pouchdb-server issue