@nename0 2017-11-23
Ok, the checking of the cursor [here](https://github.com/pouchdb/pouchdb/pull/6858#discussion-diff-152769256R177) is needed, because when allDocs is called without limit option the cursor just finishes normally with all docs. We don't have this in the `idb` adapter, because we have #6059 there.
As I already said/asked [here](https://github.com/pouchdb/pouchdb/issues/6543#issuecomment-346197236) the  optimizations in #6059 and #6033 are also suited for `indexeddb`