@nename0 2017-11-24
After rebase on #6865 the pouchdb-server build passes now. The [error](https://travis-ci.org/pouchdb/pouchdb/jobs/306757153) with AUTO_COMPACTION and phantomjs seems to be another random one.